Refactor the folders organization
