+++ app/src/main/java/com/scalar/scalar/ui/creation/CreateAccountActivity.kt
@@ -1,8 +1,8 @@
-package com.scalar.scalar.creation
+package com.scalar.scalar.ui.creation
 
 import android.content.Intent
 import android.os.Bundle
-import com.scalar.scalar.BaseActivity
+import com.scalar.scalar.ui.activities.BaseActivity
 import com.scalar.scalar.R
 import kotlinx.android.synthetic.main.activity_create_account.*
