+++ app/src/main/java/com/scalar/scalar/ui/importwallet/ImportActivity.kt
@@ -1,7 +1,7 @@
-package com.scalar.scalar.importwallet
+package com.scalar.scalar.ui.importwallet
 
 import android.os.Bundle
-import com.scalar.scalar.BaseActivity
+import com.scalar.scalar.ui.activities.BaseActivity
 import com.scalar.scalar.R
 
 class ImportActivity : BaseActivity() {
