+++ app/src/main/res/layout/activity_create_account.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".creation.CreateAccountActivity">
+    tools:context=".ui.creation.CreateAccountActivity">
 
     <TextView
         android:layout_width="wrap_content"
