+++ app/src/main/res/layout/activity_import.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".importwallet.ImportActivity">
+    tools:context=".ui.importwallet.ImportActivity">
 
     <TextView
         android:layout_width="wrap_content"
