+++ app/src/main/res/layout/activity_recovery_phrase.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".creation.RecoveryPhraseActivity">
+    tools:context=".ui.creation.RecoveryPhraseActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
