+++ app/src/main/res/layout/fragment_phrase.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".creation.PhraseFragment">
+    tools:context=".ui.creation.PhraseFragment">
 
     <TextView
         android:layout_width="wrap_content"
