+++ app/src/main/res/layout/fragment_recovery_phrase_intro.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".creation.RecoveryPhraseIntroFragment">
+    tools:context=".ui.creation.RecoveryPhraseIntroFragment">
 
     <TextView
         android:layout_width="wrap_content"
